NCBC-1777: Don’t dispatch observe ops if both ReplicateTo & PersistTo are Zero
MOTIVATION
----------
SendWithDurability(Async) always dispatches at least one observe
operation even if both ReplicateTo and PersistTo are both Zero. In this
scenario, we can determine durability has been met because the original
operation was successful.

MODIFICATIONS
-------------
- Check if both ReplicateTo and PersistTo are Zero and don't create a
  KeySequenceObserver if they are for all SendWithDurability variants
- Add unit tests to verify synchronous behaviour

RESULT
------
The SDK will not dispatch any Observe ops for requests that specify
durability requirements where both ReplicateTo and PersistTo are Zero.

NOTE: It was not possible to test the async paths because
SendWithDurabilityAsync internally calls SendWithRetryAsync and does not
have the option of providing a TaskCompletionSouce. This means it's not
possible to mock a response from the server.

Change-Id: I74187de0c151b706ec55aa336007b8a25dd1812d
Reviewed-on: http://review.couchbase.org/99948
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>