NCBC-1779: Add bucket name and config to logging for DEBUG and TRACE
Motivation
----------
Adds additional data to be logged so that it is easier to track
configuration updates. The data is missing and makes it nearly impossible
to correlate a config update to a specific bucket instance if multiple
bucket instances are used.

Modifications
-------------
 - Log the cluster map config when in TRACE mode on
   ClusterContoller.ProcsessConfig
 - Add bucket name to DEBUG logs in CarrierPublicationProvider and
   HttpStreamingProvider.

Result
------
It is now easier to correlate logs for a specific bucket instance.

Change-Id: I07276cfb252fae8e6307c343c9e6b01557718aea
Reviewed-on: http://review.couchbase.org/99957
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>