NCBC-1782: Fix unit test that checks for ServiceNotSupportedException
MOTIVATION
----------
There is a unit test that checks when the key service is not available
and SendWithRetry is used on CouchbaseRequestExecutor, the thrown
exception is ServiceNotSupportedException. However, it is possible for
an AggregateException to be thrown (multiple exceptions) and then for
one of the InnerExceptions to be a ServiceNotSupportedException.

MODIFICATIONS
-------------
- update unit test to catch both ServiceNotSupportedException or
  AggregateException and then check for ServiceNotSupported in the inner
  exceptions

RESULT
------
When the test runs, it checks for both ServiceNotSupoortedExeption or
AggregateException and then checks the inner list of exceptions.

Change-Id: Ic34e6642c246a595af46b1b5c0a4b67b6958a2b4
Reviewed-on: http://review.couchbase.org/100038
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Build Bot <build@couchbase.com>