NCBC-1791: Make sure kv service is only enabled if in nodes list
Motivation
----------
The nodesExt list may contain a local resource that is not ready for use
yet, like KV. When K/V serveice is found in the nodes list, it is ready
for use. See JVMCBC-564 for more details.

Modifications
-------------
 - Add case where both node and nodesExt list occur in cluster map
 - Ensure that node list overrides nodesExt list when determining if a
   node can be used for K/V
 - Add unit tests
 - Add example cluster map json

Result
------
If nodes and nodesExt exist, the client will use nodes to tell if KV is
enabled.

Change-Id: Ie7da32fec36bc0c4592951ae8cd09d59ace901c8
Reviewed-on: http://review.couchbase.org/100478
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>