NCBC-1791: Make sure kv service is only enabled if in nodes list
Motivation
----------
The nodesExt list may contain a local resource that is not ready for use
yet, like KV. When K/V service is found in the nodes list, it is ready
for use. See JVMCBC-564 for more details. Also fixes NCBC-1795 which is a
regression caused by NCBC-1791.

Modifications
-------------
 - If K/V is available, recreate the node and add the K/V service
 - Add test to ensure that the context is properly using the latest config
 - Add case where both node and nodesExt list occur in cluster map
 - Add configs.json which has a list of configs illustrating the scenario
 - Ensure that node list overrides nodesExt list when determining if a node can
 be used for K/V
 - Add unit tests for changes made in BucketConfigExtensions
 - Add example cluster map json

Result
------
If nodes and nodesExt exist, the client will use nodes to tell if KV is
enabled.

Change-Id: I7785d4a51909e884d3c948e0bf0237a3012b933f
Reviewed-on: http://review.couchbase.org/100942
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>