NCBC-1783: ClusterController.ProcessConfig throws System.ArgumentNullException
Motivation
----------
This is an error in Core where the internal semaphore is set to null when
Dispose is called on the BlockingCollection. When Dispose is called by the
client the client correctly stops accepting configs by calling
CompleteAdding on the blocking collection and then calls Dispose after
removing any items that may be in enqueued. This is on the dedicated
config thread (CT); however, worker threads from operation requests still
firing may try to enqueue. Since the object is disposed (and should throw
an ODE) it throws a ArgumentNullException because the internal semaphore
is already null.

Modifications
-------------
 - Suppress ArgumentNullException if it's from a Semaphore
 - Check if disposed before allowing a thread to add to the blocking
   collection

Result
------
ArgumentNullException's will be ignored if they are from the blocking
collection and are from a semaphore.

Change-Id: I851b28e7305d72505fdd9882c9ee3b5f8888aef6
Reviewed-on: http://review.couchbase.org/101197
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>