NCBC-1785: MultiplexingConnection throws ObjectDisposedException during Send
Motivation
----------
This is a side effect of allowing more than one thread to use the socket -
by design in MUX - the reading thread determined that the connection had
been terminated or otherwise failed and a worker thread was sending
another operation concurrently. In this case the operation being sent will
trigger this exception and it will be logged and prepared for retry if
applicable on a new connection or a different node depending upon the
situation.

Modifications
-------------
 - Suppress logging of ObjectDisposedExceptions in HandleDisconnect - the
   connection has already been disposed.

Result
------
ODE should not be logged for this situation.

Change-Id: I9af802b166592fd4e85064ea314273170461ac05
Reviewed-on: http://review.couchbase.org/101199
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>