NCBC-1794: Add experimental support for deferred Analytics queries
MOTIVATION
----------
Couchbase Server 6.0 introduces a new asynchronous method for executing
long running Analytics queries. These deferred queries return a handle
back that is a URI that can be polled to get the executing query's
status and results.

NOTE: This is an experimental feature and may change in the future.

MODIFICATIONS
-------------
- remove the experimental ExecutionMode analytics query parameter
- add deferred(bool) method and IsDeferred property analytics query
  parameter
- update analytics client to indicate query success if the query status
  is either Success or Running
- update analytics client to pass the http client and data mapper into
  the utility class that maps the results for the handle to use
- add IAnalyticsDeferredResultHandle and AnalyticsDeferredResultHandle
  to manage collecting the handle URI and exposing methods to get the
  query's current status and the results
- add supporting unit and integration tests

RESULT
------
It is not possible to be able to create and submit an analytics query in
the new deferred way. The result of the query can then poll the query
status and retrieve the query results once available.

Change-Id: I1208723c45fa97849900833b5c041ebe9819e78c
Reviewed-on: http://review.couchbase.org/101235
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>