NCBC-1784: Improve TryTake schmetics in ThresholdLoggerTracer
MOTIVATION
----------
The blocking collection used with ThresholdLoggingTracer can throw a
Null Argument Exception if the TryTake method is executed after the
queue has been disposed.

MODIFICATIONS
-------------
- check the collection is still active and has items before trying to
  take an item from blocking collection
- catch ArgumentNullException and check the message to see if it
  contains SemaphoreSlim, log error as normal if it doesn't

RESULT
------
The blocking collection used within ThresholdLoggingTracer should now
only try to take an item from the collection if the collection is active
and has an item to take.

If a ArgumentNullException is thrown from the SemaphoreSlim used within
the blocking colleciton, no log message is set.

Change-Id: Iaf007ac6c7df2ac600f0c6ce023d68400eea88c9
Reviewed-on: http://review.couchbase.org/101338
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>