NCBC-1808: Disable KV service in NodeAdapter if Node is null
MOTIVATION
----------
When creating a NodeAdapter, both a node and NodeExt are passed as
constructor arguments. Port mappings should be preferred from the
NodeExt, however, it is possible that the KV service has not yet fully
initialised and the node argument is null. In this scenario, the KV
service should be marked as disabled to prevent operations being
dispatched to it when the service is unavailable. An updated bucket
config will be sent once the service is available.

In addition, a bug where when comparing NodeAdapater collections using
NodeAdapaterExtensions.AreEqual was found that does not compare the
adapter's KV port. This is required because the Server list in
ConfigContxtBase uses a dictionary of Endpoints which includes the port
and should be considered different if the port is different.

MODIFICATIONS
-------------
- when creating a NodeAdapter using a NodeExt, check if the Node was
  null and disable KV(SSL) by settings adapter ports to 0
- add unit test to verify behaviour for NodeAdapater with null Node

- fix bug in NodeAdapterExtensions where the node adapter's KV port is
  not compared in AreEqual
- move NodeAdapterExtensions tests from Couchbase.Tests to
  Couchbase.UnitTests including associated json config files
- add unit test to verify KV port is part of the AreEqual comparison
  when comparing NodeAdapater lists

RESULT
------
When building a NodeAdapater, the KV service is disabled if a Node
structure was passed in and the KV port is set to 0. Also, when
comparing node adapter lists via NodeAdapaterExtensions, adapters with
different ports are not considered equal.

Change-Id: I60369618dc4ed6f4f1bc82e6fc58d8634f5ce1fc
Reviewed-on: http://review.couchbase.org/101505
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>