NCBC-1824: Separate version prefix and suffix in csproj
MOTIVATION
----------
When building the project, it is desirable to be able to independently
set the version prefix (semver number) and the suffix (revision). This
would be especially useful during CI where the rev can be the CI build
number and then omitted for a release to create a releasable package.

MODIFICATIONS
-------------
- separate version prefix and suffix in Couchbase.csproj

RESULT
------
The version suffix can now be set independently of the version prefix
using dotnet pack.

Change-Id: I05233029723beaf57d9f2afea595c25b2c496671
Reviewed-on: http://review.couchbase.org/102530
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>