NCBC-1813: Reuse cluster ref between tests
MOTIVATION
----------
It can take a long time to setup a cluster and open a bucket when
running against CouchbaseMock, instead it would be more desirable to
create a single instance and reuse within the test case.

MODIFICATIONS
-------------
- establish a cluster and bucket reference once and reuse through tests
  in both KV and subdoc integration tests

RESULT
------
A single cluster reference is used for all tests in each test class to
speed up tests.

Change-Id: I660e8b161f2e2801b0e5508db368eb2a58dccfe6
Reviewed-on: http://review.couchbase.org/102653
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>