NCBC-1825: Replace AppVeyor build config with Jenkins
MOTIVATION
----------
With the new build infrastructure, we no longer require the build-utils
scripts & files and the appveyor configuration, these can all be
removed. Also, we can replace the AppVeyor build badge with the Jenkins
version.

MODIFICATIONS
-------------
- remove build-utils folder and contents
- remove appveyor.yml
- replace appveyor build badge with Jenkins version in README

RESULT
------
The build-utils and appveyor config are no longer part of the repository
and we use the Jenkins build badge.

Change-Id: Iaaae383bd9e3d788b565eb2de22de063053b629e
Reviewed-on: http://review.couchbase.org/102918
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>