NCBC-1829: Add assembly optimizations when building the SDK in Release
MOTIVATION
----------
When building in Release configuration, it is desirable to optimize the
assembly and not embed debug symbols to maximize performance. However,
the new csproj format does not automatically set these properties and
they need to be set manually.

MODIFICATIONS
-------------
- add configuration to csproj to set optimize = true and
  embedDebugSymbols = false for Release configuration

RESULT
------
When building the Couchbase SDK in Release configuration, the assembly
has the optimized property set and doesn't embed debug symbols.

Change-Id: I0e806c516c4cccf21fdb41b9036706476a4d136f
Reviewed-on: http://review.couchbase.org/103554
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>