Add new license SPDX license identifier for Apache 2.0
Motivation
----------
The license element <LicenseUrl/> has been deprecated in Nuget, so we need
to update to use a SPDX license identifier element instead for an Apache
2.0 license.

Modifications
-------------
 - Add <PackageLicenseExpression/> and SPDX identifier for Apache 2.0 license
to csproj.

Result
------
We no longer use a url that points to the Apache 2.0 license; we use an
SPDX expression which points to a Apache 2.0 license on Github.

Change-Id: I84632da482753fb7b0a88d975e4ebd1aa7d1dd38
Reviewed-on: http://review.couchbase.org/103569
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Mike Goldsmith <goldsmith.mike@gmail.com>