NCBC-1833: Update Error Handling for SDK3
Motivation
----------
Error handling has changed in SDK 3.0; we no longer return the Status
and Exception in  the response object, but instead throw exceptions for
basic K/V and other situations. The caller must handle the error by
catching the thrown exception and reacting according to the situation.

Modifications
-------------
 - Add various exception classes
 - Update K/V to throw KeyValueException with status
 - Add unit tests

Result
------
When an error occurs an exception is raised.

Change-Id: I89dddf2a519fce8d3f3568f0c17c5ef954256365
Reviewed-on: http://review.couchbase.org/106021
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>