NCBC-1885: Add Span<T> implementations to IByteConverter
Motivation
----------
Allow callers using more efficient Memory<T> and Span<T> structures to
use IByteConverter, rather than being limited to callers using byte[].

Modifications
-------------
Added IByteConverter methods for ToXXX that accept ReadOnlySpan<T>,
and IByteConverter methods for FromXXX that accept Span<T>. However,
for methods that have an overload with `useNbo`, only implement that
overload on IByteConverter.

Implemented overloads without `useNbo` as extension methods in
ByteConverterExtensions. For the FromXXX methods where the byte[] form
was taking a reference and could create a new byte array, added an
extension method that returns a new byte[] instead.

Added `GetStringByteCount` to the interface to provide a consistent
point to call to get the number of bytes for an encoded string, so
callers don't need to know that the internal implementation is UTF-8.

Refactored all byte[] methods to call one of the new methods that
accepts a span.

Results
-------
The public IByteConverter API now supports all operations via a span.
The exception is calls which expect a new array to be allocated, which
now has a less confusing syntax but still returns a byte array for
efficiency.

Current byte[] operations all forward their work to span based
implementations, so the byte array unit tests are effectively testing
both implementations.

The byte array overloads will be gradually phased out and removed.

Change-Id: I245cc782570878ebfa4760811ae4db275df88716
Reviewed-on: http://review.couchbase.org/106670
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>