NCBC-1896: Use shared memory for operation read
Motivation
----------
Reduce heap allocations, especially on the LOH, and reduce bitblitting.

Modifications
-------------
Created SlicedMemoryOwner, with associated tests, to support using
length-limited buffers from the shared memory pool.

Made IOperation return Data as a Memory<byte>, accept data in ReadAsync
as IMemoryOwner<byte>, and inherit IDisposable so that memory is
released when the operation is disposed.

Updated usages of IOperation.Data based on the new type, except for
the creation of results.

Update usage of the operations to dispose of them after completion.

Updated data flow from MultiplexingConnection all the way to
IOperation.ReadAsync to pass data using IMemoryOwner<byte>, assuming
that each step in the path is moving ownership to the next step.

Results
-------
With fewer intermediate conversions between byte[] and MemoryStream
from MultiplexingConnection all the way through operation parsing,
this will reduce bit blitting. By using a shared memory pool rather than
new byte arrays, heap allocations and garbage collection are reduced.
This should be particularly noticable when dealing with responses
between 80KB and 1MB in size.

The risk with these changes is improper handling of IMemoryOwner<byte>
semantics around the concept of ownership and ensuring Dispose is
called. Not doing so will result in memory leaks. Information about the
semantics, as described by Microsoft, is available at
https://docs.microsoft.com/en-us/dotnet/standard/memory-and-spans/memory-t-usage-guidelines.

Note: There is still additional bit blitting and array creation
surrounding movement of the data from the operation to the result
objects. This will be addressed in a future commit.

Change-Id: I2618525dab258cc09826a785395237927a45b03a
Reviewed-on: http://review.couchbase.org/106834
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>