NCBC-1898: Create load test framework
Motivation
----------
Make memory and CPU profiling easier with a set of tests which run load
through the SDK.

Modifications
-------------
Created Couchbase.LoadTests assembly, with some parts modelled loosely
on cbc-pillowfight.

Results
-------
Tests are available which run bulk operations with a set of small
documents or a set of large documents.

Change-Id: If3152774f927d41db81a74ec60407d31726c4ba1
Reviewed-on: http://review.couchbase.org/106929
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>