NCBC-1897: Use shared memory for get results
Motivation
----------
Reduce heap allocations, especially on the LOH, and reduce bitblitting.

Modifications
-------------
Copy data to GetResult and LookupInResult using IMemoryOwner<byte>,
transferring ownership to the result, rather than making a new byte
array for the result.

Add IDisposable to IGetResult so that memory may be freed by the
consumer. Use finalizers to protect against consumers forgetting to
call dispose.

Change OperationSpec to keep a slice of the memory buffer, rather than
yet another byte array copy.

Update tests to dispose of results.

Results
-------
A further reduction in heap allocations and bitblitting related to
reading results from K/V operations.

Note: Legacy sub document operations (not using LookupInResult) still
use a byte array copy for their OperationSpecs. This is because they
could be accessed after the operation, and there for the shared memory,
is disposed. It is assumed that these operations will be deprecated.

Change-Id: I0be66f8b8efff673c0c9b8167b48e79523a1c6fa
Reviewed-on: http://review.couchbase.org/107038
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>