NCBC-1906: Analytics URI cache is empty
Motivation
----------
When a swap/failover and rebalance is triggered, the cached URIs for the
analytics service may all be set to failed. This commit clears the
failed state so the URI can be retried in case it was a temp failure.

Modifications
-------------
 - Make ConfigContextBase.GetAnalyticsUri reset the URIs when all are in
 a failed state so that the they can be tried again.

Result
------
URIs should be always be available whether or not they succeed initially
or not - this is the same behavior as Views, N1QL and FTS.

Change-Id: I4b9f1d1063a0f6c34d7844f9b52556a3c2c818aa
Reviewed-on: http://review.couchbase.org/107262
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>