NCBC-1872: Move ScanConsistency to root namespace
MOTIVATION
----------
The ScanConsistency enum is used in a couple of different services, but
the enum currently exists in the Couchbase.Services.Query namespace.
This should be moved to a more generic location.

MODIFICATIONS
-------------
- move ScanConsistency to root Couchbase namespace

RESULT
------
The ScanConsistency enum is now in a more appropriate location as it's
used by a number of services.

Change-Id: Ic941f543377473d68fc383159be6d80a280cb99e
Reviewed-on: http://review.couchbase.org/107540
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>