NCBC-1911: Write keys to OperationBuilder
Motivation
----------
Improve efficiency of extras generation by reducing heap allocations.

Modifications
-------------
Replaced CreateKey with WriteKey, which accepts an OperationBuilder.
Uses a memory pool block instead of a heap allocated array to build the
key. Also reduces operations by not precalculating the size beforehand,
instead allocating enough for the maximum LEB128 and maximum key, and
slicing the memory to the actual length.

Results
-------
Reduced heap allocations and overall operations when writing the key.
There is still an extra bitblit taking place to the shared memory pool.
This will be addressed on .NET Core 2.1 with a future improvement to
RecyclableMemoryStream.

Change-Id: Ia76d4a47f2ff1bf27c0e1636b12db30e3c9336ef
Reviewed-on: http://review.couchbase.org/107481
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>