NCBC-1914: Write body to OperationBuilder
Motivation
----------
Improve efficiency of extras generation by reducing heap allocations.

Modifications
-------------
Replaced CreateBody with WriteBody, which accepts an OperationBuilder.
For many operations, uses a stack allocated span instead of a heap
allocated array to build the body. For some subdoc operations, reduces
CPU operations by not precalculating the path size beforehand, instead
allocating enough for the maximum length and slicing the memory to the
actual length.

Results
-------
Reduced heap allocations and overall operations when writing the body.
There is still an extra bitblit taking place to the shared memory pool.
This will be addressed on .NET Core 2.1 with a future improvement to
RecyclableMemoryStream.

This doesn't yet address the heap allocations involved in serialization
of the body, this work will come in a future commit.

Change-Id: I9a927b90702e6e781a313d34e994781a3f7c8fe1
Reviewed-on: http://review.couchbase.org/107530
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>