NCBC-1901: Encode/serialize using streams
Motivation
----------
Reduce bitblitting and heap allocations by serializing directly onto
the RecyclableMemoryStream for the operation, rather than an
an intermediate memory stream and an intermediate byte array.

Modifications
-------------
Refactored ITypeTranscoder and DefaultTranscoder to accept a Stream
rather than return a byte array.

Refactored ITypeSerializer and DefaultSerializer to accept a Stream
rather than return a byte array. Add an extension method that returns a
byte array to allow fallback to the old behavior for consumers.

Refactored operations to use the new API surface.

Switch to using MemoryStreamFactory for handling creation of
RecyclableMemoryStream so that it can share the memory manager between
OperationBuilder and subdoc operations.

Refactored unit tests fo rthe new API surface.

Results
-------
Serialization is much more efficient for most cases. For array subdoc
operations with StripBrackets == true there is still an intermediate
array required, but it now uses RecyclableMemoryStream for efficiency.

MemoryStreamFactory is now internal because it was necessary to return
a RecyclableMemoryStream instead of a MemoryStream so that GetBuffer
works correctly. This will allow us to replace RecyclableMemoryStream in
the future, if desired.

Change-Id: I1c9162ccfdc0fd594e98f4bd913d2cf29338cd6e
Reviewed-on: http://review.couchbase.org/107591
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>