NCBC-1921: Optimize for MemoryPool Stream Writing
Motivation
----------
Avoid extra bitblitting on .NET full and .NET Core 2.0 when writing to
streams from MemoryPool.

Modifications
-------------
Added a stream extension method that recognizes the use of array-backed
ReadOnlyMemory<T> and optimizes for it when writing.

Updated several operations to use this optimization, and also rework
DefaultTranscoder to use it as well instead of a local copy.

Changed StripBrackets to use ReadOnlyMemory instead of ReadOnlySpan so
that this optimization could be used on subdoc mutations.

Results
-------
Avoids an ArrayPool allocation and bitblit in several cases when
running on older .NET platforms where Stream doesn't implement a
ReadOnlySpan overload.

Change-Id: I582c2e73a9234ba2bb63e5015387d242d3d8cd8f
Reviewed-on: http://review.couchbase.org/107728
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>