NCBC-1875: Add support for connection strings to bootstrap Cluster
MOTIVATION
----------
SDK 3.0 should allow a Cluster object to be created and configured using
a connection string. The connection string will may optionally include
the scheme (http, couchbase, couchbases), a username, a list of hosts
with ports and a list of parameters as key-value pairs.

MODIFICATIONS
-------------
- add ConnectionString class that can parse a provided connection string
  to retrieve the
- add unit tests to verify connection strings are parsed correctly
- add constructor overload to Cluster that takes a string that is parsed
  with ConnectionString and then creates a Configuration object

RESULT
------
It's not possible to bootstrap a Cluster object using a connection
string to provide the scheme and hosts.

NOTE: In a future commit we will expand connection strings to set
configuration properties.

Change-Id: I9248d282bf5c6cbc754dab49e0246093eb30d5bf
Reviewed-on: http://review.couchbase.org/108346
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>