NCBC-1962: Reimplement OperationBuilder pooling
Motivation
----------
Reduce heap allocations and other expense related to instantiating an
OperationBuilder for each operation write. However, this time avoid
problems with bootstrapping.

Modifications
-------------
Implemented OperationBuilderPool, and updated OperationBase to use this
pool rather than a new instance on each call to WriteAsync.

Added Reset, a property to see the current Capacity, and the ability to
change the Converter to OperationBuilder.

Results
-------
4-5% improvement in the speed of WriteAsync across a variety of
document sizes (512 bytes to 512 kilobytes).

Note: To reduce memory pressure, there is a limit on the maximum
size of OperationBuilder to retain and the maximum number of builders to
retain. These settings may deserve configuration knobs in the future.

Change-Id: Iabefdb4addee103a655f6a80c4a5ff0686d8cd4e
Reviewed-on: http://review.couchbase.org/109260
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>