NCBC-1953: OperationBuilder pooling
Motivation
----------
Reduce heap allocations and other expense related to instantiating an
OperationBuilder for each operation write.

Modifications
-------------
Implemented OperationBuilderPool, and updated OperationBase to use this
pool rather than a new instance on each call to WriteAsync.

Added Reset, a property to see the current Capacity, and the ability to
change the Converter to OperationBuilder.

Results
-------
4-5% improvement in the speed of WriteAsync across a variety of
document sizes (512 bytes to 512 kilobytes).

Note: To reduce memory pressure, there is a limit on the maximum
size of OperationBuilder to retain and the maximum number of builders to
retain. These settings may deserve configuration knobs in the future.

Change-Id: Iecefa9e2fdb212df024120ff55ad995d8e486f8c
Reviewed-on: http://review.couchbase.org/108962
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>