NCBC-1944: Add cluster config manager component
Motivation
----------
The role of this component is to ensure that the latest config cluster is
always being used by the client. It recieves configs from any thread and
checks to see if the revision is higher than the previous and that the config
itself has a meaningful change (VBuckets changed or nodes) and then initiates
any topology changes to the current context.

Note that this is partial commit in that it fits within a larger scope
of bootstrapping. Its currently not 100% working, but should be isolated
from other commits.

Modifications
-------------
 - Added a "context" class for globally handling config and ensuring
 that the latest revision is being used.
 - Added IEquatable impl for all cluster map classes.
 - Added an exception that is raised when the context is a stopped state
 so that an ObjectDisposedException is not raised instead.
 - Renamed IBucketSender to IBucketInternal and added event handler for
 BucketConfigEventArgs
 - Added tests for basic cases

Result
------
When enabled in a later commit, will handle new cluster maps and check
to see if a topology change must be done.

Change-Id: Ib2f5282d070b11aebb4dfc4d367307b9684d7142
Reviewed-on: http://review.couchbase.org/108986
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>