NCBC-1955: Extension methods on primary interfaces
Motivation
----------
Using extension methods for interface overloads makes mocking the
interfaces for unit testing much easier. The consumer only needs to mock
the most powerful overload, and the extension methods fill defaults and
forward the request on to that overload. It also simplifies the surface
area of the interface and reduces clutter on the implementation.

This pattern is also consistent with ASP.NET Core practices, which will
familiar to many users.

Modifications
-------------
Move all basic overloads on ICluster, IBucket, IScope, ICollection, and
IBinaryCollection to extension methods, moving their implemenations
from the current class implementations.

Renamed some async methods to have an Async suffix on the method name
for consistency.

Results
-------
It will now be much easier for consumers to mock Couchbase when
creating unit tests for their data access layer.

Change-Id: I3584e6a80f238bdc4fde7334bf41c545f3f0d8f5
Reviewed-on: http://review.couchbase.org/109008
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>