NCBC-1957: Release connection acquired during PooledIOService constructor
MOTIVATION
----------
When creating an instance of PooledIOService, a connection from the
ConnectionPool parameter is used to check the server for available
features. However, the connection is not returned back to the pool which
means it can not be used again and another connection has to be created
in the pool.

MODIFICATIONS
-------------
- release the connection at the end of PooledIOService constructor
- add unit test to verify the connection has been released properly

RESULT
------
The connection used during PooledIOService construction is released back
to the pool and can be re-used. Also, an additional connection is not
created to replace the original one.

Change-Id: I54a3140aa16b8173934732e51995a402af89f500
Reviewed-on: http://review.couchbase.org/109049
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>