NCBC-1927: Resolve vbucket server map endpoints if different from bucket config server list
MOTIVATION
----------
When bootstraping the client receives a vbucket server map from as part
of the bucket config. This server map is then used to map active and
replica keys to cluster nodes. However, if the nodes are configured
using an ingress system (eg Kubernetes) the nodes may not be reachable
on the endpoints contained in the bucket config.

MODIFICATIONS
-------------
- during CCCP bootstrap, if the list of node hostnames is different to
  the resolved IPs for the nodes, the client uses the IPs instead
- add method to clear vbucket server list IPs

RESULT
------
The SDK can now successfully connect to a Couchbase cluster when the
bucket config's nodes are internal (eg public Kubernetes cluster) by
using resolved IP Endpoints instead.

Change-Id: I3ff4f5c013f8bc1955603e9f00243f1ce05356bd
Reviewed-on: http://review.couchbase.org/109050
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>