NCBC-1816: Add support for exporting/importing deferred Analytics queries
MOTIVATION
----------
As part of an up coming Couchbase Server release, it will be possible to
created deferred Analytics queries where the server will return a handle
that can be used to query the status and retrieve results. The SDKs need
to be able to allow these handles be persisted between SDK life cycles
with an Export/Import system.

MODIFICATIONS
-------------
- add Export/Import methods to IBucket, CouchbaseBucket and
  MemcachedBucket, IAnalyticsClient and AnalyticsClient
- add unit tests to verify expected behaviour

RESULT
------
It's now possible to export and import deferred analytics query handles
using the SDK.

Change-Id: I59d64c730e3146d2e01b13f9f65812dc88f60858
Reviewed-on: http://review.couchbase.org/109461
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>