NCBC-1984: Timestamp underflow causes expiration to be set to infinite
Motivation
----------
Fixes a casting issue that happens during the unix time conversion - at a point the expiry is converted to a TimeSpan and then the TotalMilliseconds property is used which converts the value to a double (.999) which then is converted to a unsigned integer thus .NET makes it a zero.

Modifications
-------------
 - Ensure values less than 1000ms are converted to 1s. All other
 non-whole seconds are truncated down to the next whole number; e.g.
 2500ms is rounded down to 2s.
 - Add unit test

Results
-------
When the expiry is set as an unsigned int, this value is passed directly
to IO without converting to a TimeStamp first.

Change-Id: If0ec9914f8088776f3b82fa77eab198a1eafeda5
Reviewed-on: http://review.couchbase.org/110385
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>