NCBC-1986: Enable cluster map update changes to refresh client state
Motivation
----------
When a cluster map config arrives, check to see if it has modifications.
If it has modifications, then either update the VBucketMap if it's
changed or the internal connections if the topology has changed.

Modifications
-------------
 - Add logic for handling cluster configs
 - Add code to detect what has changed
 - Update the internal state of the Bucket if its changed

Results
-------
A change in the cluster map should be reflected in the client's state.

Change-Id: Ic35e210b917b7a153a6fdfee874f2e9cba5b10b4
Reviewed-on: http://review.couchbase.org/110603
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>