NCBC-2002: ServiceNotSupportedException: Data service not found
Motivation
----------
Fixes a regression caused by NCBC-1927 where if an alternate address may
cause the data service to appear to not be configured. Alternate address
is encountered in environments like Kubernetes where the external
address and the internal addresses differ.

Modifications
------------
 - Check for alternate address; if found make the SDK use the alternate
 address for all internal communication
 - Remove logic introduced in NCBC-1927
 - Remove OnDerserialization event from VBucketKeyMapper and instead
 delagate call in OnDerserilization event in BucketConfig

Result
------
If alternate addresses are detected the client will use them; the client
will detect the data service and not error.

Change-Id: Ie7cacce15c8fc725e42dc3c8f6b89ad0be06a6a9
Reviewed-on: http://review.couchbase.org/111462
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>