NCBC-2021: Update MultiMutation to correctly read mutation token and offsets
MOTIVATION
----------
When executing a Subdoc multi-mutation that includes a counter op that
returns a value with enhanced durability enabled, the mutation result
is incorrectly read because the body offsets are not correctly handled.

MODIFICATIONS
-------------
- set OperationBase._bodyLength when writing the request body
- set MutationToken after reading body
- use BodyOffset not Extras offset to read response body
- add unit tests to verify expected behaviour for both sync and async
  paths

RESULT
------
SubDoc multimutations correctly read mutation tokens and manage body
offsets when using a counter operation and enhanced durability is
enabled.

Change-Id: I07096d56182d5ba8790cd73af3f17ccb2d61d222
Reviewed-on: http://review.couchbase.org/111743
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>