NCBC-2028: Stream.Timeout cannot be set on all platforms
Motivation
----------
This fixes a regression in NCBC-2025 where the Stream.Timeout is setting in the HTTP Streaming mechanism, however, this is not supported on all platforms and causes an exception if Stream.CanTimeout is false and Steam.Timeout is set.

Modifications
-------------
 - Add conditional check to ensure Stream.CanTimeout is true before
 setting Stream.Timeout to infinite in ConfigThreadState.cs

Result
------
If the platform does not support setting Stream.Timeout, then it will
not be set to Timeout.Infinite.

Change-Id: I92db4a1768b7360e35d7be8a338dd0a0306a656e
Reviewed-on: http://review.couchbase.org/111934
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>